# net, config: optimize read and write packets