<!--

Thank you for contributing to TiProxy!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and 
linking the relevant issues via the "close: #xxx" or "ref: #xxx".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #381 

Problem Summary:
`WritePacket` and `ReadPacket` become the hot path when there are many rows returned, so this code should be optimized:
- `WriteOnePacket` calls 2 `NewReader` and that allocates memory twice.
- `ReadPacket` grows slice in `data = append(data, buf)`, which is unncessary.
- `var header [4]byte` escapes to the heap because `header` is passed as a parameter to other functions.
- The read/write buffer size is 16K, which means for 1000 rows, it calls syscall `read`/`write` 8 times.
- `io.ReadFull` contains unnecessary boundary checks, function calls, and interface conversion.

What is changed and how it works:
Optimization:
- Replace `io.Copy` with `readWriter.Write` to remove `NewReader`.
- Do not call `data = append(data, buf)` for the first packet.
- Move `header` into `PacketIO`/`compressedReadWriter` as a member.
- Make `conn-buffer-size` configurable and set to 32K by default.
- Replace `io.ReadFull` with customized `ReadFull`.

Others:
- Add `DirectWrite` to `compressedReadWriter` because it was missed, but it doesn't matter actually.
- Also add benchmarks of compression.

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [x] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No code


Sysbench result:
Before this PR:
| Range size | QPS   | TiProxy CPU | QPS per 100% CPU |
|------------|-------|-------------|------------------|
| 10         | 30955 | 180%        | 17200            |
| 100        | 14655 | 190%        | 7710             |
| 1000       | 2112  | 200%        | 1060             |
| 10000      | 221   | 200%        | 110              |

After this PR:
| Range size | QPS   | TiProxy CPU | QPS per 100% CPU |
|------------|-------|-------------|------------------|
| 10         | 32818 | 150%        | 21878            |
| 100        | 22392 | 190%        | 11785            |
| 1000       | 4093  | 200%        | 2046             |
| 10000      | 449   | 200%        | 224              |

The flame graph when range size is 1000:

![image](https://github.com/pingcap/tiproxy/assets/29590578/a4f87e2f-7704-4f3b-9766-08490bd8953a)


Notable changes

- [x] Has configuration change
- [ ] Has HTTP API interfaces change
- [ ] Has tiproxyctl change
- [ ] Other user behavior changes

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```